Update Command Names
