+++ src/main/java/frc/robot/commands/AimNote.java
@@ -4,69 +4,67 @@
 
 package frc.robot.commands;
 
+import java.util.function.BooleanSupplier;
+import java.util.function.DoubleSupplier;
+
 import edu.wpi.first.math.filter.SlewRateLimiter;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.Constants;
 import frc.robot.Constants.OperatorConstants;
+import frc.robot.subsystems.LimeLightSubsystem;
 import frc.robot.subsystems.SwerveSubsystem;
-import java.util.function.BooleanSupplier;
-import java.util.function.DoubleSupplier;
 
-public class ManualDriveCommand extends Command {
-  /** Creates a new ManualDriveCommand. */
+public class AimNote extends Command {
+  /** Creates a new AimNoteCommand. */
+  private final LimeLightSubsystem m_limeLightSubsystem;
   private final SwerveSubsystem m_swerveSubsystem;
   // Inputs
   private DoubleSupplier xSpeedFunc;
-  private DoubleSupplier ySpeedFunc;
-  private DoubleSupplier zSpeedFunc;
   private BooleanSupplier isSlowModeFunc;
   // Slew rate limiter
   private final SlewRateLimiter xLimiter;
-  private final SlewRateLimiter yLimiter;
-  private final SlewRateLimiter zLimiter;
-
-  public ManualDriveCommand(SwerveSubsystem swerveSubsystem, DoubleSupplier xSpeed, DoubleSupplier ySpeed, DoubleSupplier zSpeed, BooleanSupplier isSlowMode) {
+  public AimNote(LimeLightSubsystem limeLightSubsystem, SwerveSubsystem swerveSubsystem, DoubleSupplier xSpeed, BooleanSupplier isSlowMode) {
+    // Use addRequirements() here to declare subsystem dependencies.
+    this.m_limeLightSubsystem = limeLightSubsystem;
     this.m_swerveSubsystem = swerveSubsystem;
     this.xSpeedFunc = xSpeed;
-    this.ySpeedFunc = ySpeed;
-    this.zSpeedFunc = zSpeed;
     this.isSlowModeFunc = isSlowMode;
     xLimiter = new SlewRateLimiter(4);
-    yLimiter = new SlewRateLimiter(4);
-    zLimiter = new SlewRateLimiter(4);
-    addRequirements(m_swerveSubsystem); 
+    addRequirements(m_limeLightSubsystem, m_swerveSubsystem);
   }
 
+  // Called when the command is initially scheduled.
+  @Override
+  public void initialize() {}
+
+  // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    // Get value
+        // Get value
     // 負號加在這
     double xSpeed = -xSpeedFunc.getAsDouble();
-    double ySpeed = -ySpeedFunc.getAsDouble();  
-    double zSpeed = -zSpeedFunc.getAsDouble();
+    double zSpeed = m_limeLightSubsystem.getTurnPIDOutput();
     boolean isSlowMode = isSlowModeFunc.getAsBoolean();
     // Dead band Limit
-    xSpeed = Constants.DeadBandLimit(xSpeed, OperatorConstants.kJoystickDeadBand);
-    ySpeed = Constants.DeadBandLimit(ySpeed, OperatorConstants.kJoystickDeadBand);
-    zSpeed = Constants.DeadBandLimit(zSpeed, OperatorConstants.kJoystickDeadBand);
+    Constants.DeadBandLimit(xSpeed, OperatorConstants.kJoystickDeadBand);
     // TurboModeSelect
     if(isSlowMode){
       xSpeed = xSpeed*0.4;
-      ySpeed = ySpeed*0.4;
-      zSpeed = zSpeed*0.4;
+      zSpeed = zSpeed*1;
     }
     else{
-      xSpeed = xSpeed*1;
-      ySpeed = ySpeed*1;
+      xSpeed = xSpeed*0.85;
       zSpeed = zSpeed*1;
     }
     // SlewRate
     xSpeed = xLimiter.calculate(xSpeed);
-    ySpeed = yLimiter.calculate(ySpeed);
-    zSpeed = zLimiter.calculate(zSpeed);
     // Output
-    m_swerveSubsystem.drive(xSpeed, ySpeed, zSpeed, true);
+    m_swerveSubsystem.drive(xSpeed, 0, zSpeed, false);
   }
+
+  // Called once the command ends or is interrupted.
+  @Override
+  public void end(boolean interrupted) {}
 
   // Returns true when the command should end.
   @Override
