+++ src/main/java/frc/robot/commands/ManualDrive.java
@@ -4,67 +4,69 @@
 
 package frc.robot.commands;
 
-import java.util.function.BooleanSupplier;
-import java.util.function.DoubleSupplier;
-
 import edu.wpi.first.math.filter.SlewRateLimiter;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.Constants;
 import frc.robot.Constants.OperatorConstants;
-import frc.robot.subsystems.LimeLightSubsystem;
 import frc.robot.subsystems.SwerveSubsystem;
+import java.util.function.BooleanSupplier;
+import java.util.function.DoubleSupplier;
 
-public class AimNoteCommand extends Command {
-  /** Creates a new AimNoteCommand. */
-  private final LimeLightSubsystem m_limeLightSubsystem;
+public class ManualDrive extends Command {
+  /** Creates a new ManualDriveCommand. */
   private final SwerveSubsystem m_swerveSubsystem;
   // Inputs
   private DoubleSupplier xSpeedFunc;
+  private DoubleSupplier ySpeedFunc;
+  private DoubleSupplier zSpeedFunc;
   private BooleanSupplier isSlowModeFunc;
   // Slew rate limiter
   private final SlewRateLimiter xLimiter;
-  public AimNoteCommand(LimeLightSubsystem limeLightSubsystem, SwerveSubsystem swerveSubsystem, DoubleSupplier xSpeed, BooleanSupplier isSlowMode) {
-    // Use addRequirements() here to declare subsystem dependencies.
-    this.m_limeLightSubsystem = limeLightSubsystem;
+  private final SlewRateLimiter yLimiter;
+  private final SlewRateLimiter zLimiter;
+
+  public ManualDrive(SwerveSubsystem swerveSubsystem, DoubleSupplier xSpeed, DoubleSupplier ySpeed, DoubleSupplier zSpeed, BooleanSupplier isSlowMode) {
     this.m_swerveSubsystem = swerveSubsystem;
     this.xSpeedFunc = xSpeed;
+    this.ySpeedFunc = ySpeed;
+    this.zSpeedFunc = zSpeed;
     this.isSlowModeFunc = isSlowMode;
     xLimiter = new SlewRateLimiter(4);
-    addRequirements(m_limeLightSubsystem, m_swerveSubsystem);
+    yLimiter = new SlewRateLimiter(4);
+    zLimiter = new SlewRateLimiter(4);
+    addRequirements(m_swerveSubsystem); 
   }
 
-  // Called when the command is initially scheduled.
-  @Override
-  public void initialize() {}
-
-  // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-        // Get value
+    // Get value
     // 負號加在這
     double xSpeed = -xSpeedFunc.getAsDouble();
-    double zSpeed = m_limeLightSubsystem.getTurnPIDOutput();
+    double ySpeed = -ySpeedFunc.getAsDouble();  
+    double zSpeed = -zSpeedFunc.getAsDouble();
     boolean isSlowMode = isSlowModeFunc.getAsBoolean();
     // Dead band Limit
-    Constants.DeadBandLimit(xSpeed, OperatorConstants.kJoystickDeadBand);
+    xSpeed = Constants.DeadBandLimit(xSpeed, OperatorConstants.kJoystickDeadBand);
+    ySpeed = Constants.DeadBandLimit(ySpeed, OperatorConstants.kJoystickDeadBand);
+    zSpeed = Constants.DeadBandLimit(zSpeed, OperatorConstants.kJoystickDeadBand);
     // TurboModeSelect
     if(isSlowMode){
       xSpeed = xSpeed*0.4;
-      zSpeed = zSpeed*1;
+      ySpeed = ySpeed*0.4;
+      zSpeed = zSpeed*0.4;
     }
     else{
-      xSpeed = xSpeed*0.85;
+      xSpeed = xSpeed*1;
+      ySpeed = ySpeed*1;
       zSpeed = zSpeed*1;
     }
     // SlewRate
     xSpeed = xLimiter.calculate(xSpeed);
+    ySpeed = yLimiter.calculate(ySpeed);
+    zSpeed = zLimiter.calculate(zSpeed);
     // Output
-    m_swerveSubsystem.drive(xSpeed, 0, zSpeed, false);
+    m_swerveSubsystem.drive(xSpeed, ySpeed, zSpeed, true);
   }
-
-  // Called once the command ends or is interrupted.
-  @Override
-  public void end(boolean interrupted) {}
 
   // Returns true when the command should end.
   @Override
