+++ src/main/java/frc/robot/commands/ShooterEjectNote.java
@@ -7,30 +7,29 @@ package frc.robot.commands;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.subsystems.ShooterSubsystem;
 
-public class FeedNote extends Command {
-  /** Creates a new NoteShootSpeakerCommand. */
-  private final ShooterSubsystem m_ShooterSubsystem;
-
-  public FeedNote(ShooterSubsystem shooterSubsystem) {
-    this.m_ShooterSubsystem = shooterSubsystem;
-    addRequirements(m_ShooterSubsystem);
+public class ShooterEjectNote extends Command {
+  /** Creates a new NoteBackCommand. */
+  private final ShooterSubsystem m_shooterSubsystem;
+  public ShooterEjectNote(ShooterSubsystem shooterSubsystem) {
+    // Use addRequirements() here to declare subsystem dependencies.
+    this.m_shooterSubsystem = shooterSubsystem;
+    addRequirements(m_shooterSubsystem);
   }
 
   // Called when the command is initially scheduled.
   @Override
-  public void initialize() {}
+  public void initialize() {
+    m_shooterSubsystem.Ejecting();
+  }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
-  public void execute() {
-    // 把它放在這，因為他需要時時監看是否到達目標轉速。
-    m_ShooterSubsystem.Shoot();
-  }
+  public void execute() {}
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-    m_ShooterSubsystem.StopMotors();
+    m_shooterSubsystem.StopIndexerMotor();
   }
 
   // Returns true when the command should end.
