+++ src/main/java/frc/robot/commands/ShooterFeedNote.java
@@ -7,29 +7,30 @@ package frc.robot.commands;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.subsystems.ShooterSubsystem;
 
-public class NoteBackCommand extends Command {
-  /** Creates a new NoteBackCommand. */
-  private final ShooterSubsystem m_shooterSubsystem;
-  public NoteBackCommand(ShooterSubsystem shooterSubsystem) {
-    // Use addRequirements() here to declare subsystem dependencies.
-    this.m_shooterSubsystem = shooterSubsystem;
-    addRequirements(m_shooterSubsystem);
+public class ShooterFeedNote extends Command {
+  /** Creates a new NoteShootSpeakerCommand. */
+  private final ShooterSubsystem m_ShooterSubsystem;
+
+  public ShooterFeedNote(ShooterSubsystem shooterSubsystem) {
+    this.m_ShooterSubsystem = shooterSubsystem;
+    addRequirements(m_ShooterSubsystem);
   }
 
   // Called when the command is initially scheduled.
   @Override
-  public void initialize() {
-    m_shooterSubsystem.Ejecting();
-  }
+  public void initialize() {}
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
-  public void execute() {}
+  public void execute() {
+    // 把它放在這，因為他需要時時監看是否到達目標轉速。
+    m_ShooterSubsystem.Shoot();
+  }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-    m_shooterSubsystem.StopIndexerMotor();
+    m_ShooterSubsystem.StopMotors();
   }
 
   // Returns true when the command should end.
